+++ src/main/java/org/betonquest/betonquest/compatibility/redischat/RedisChatInterceptor.java
@@ -1,6 +1,7 @@
 package org.betonquest.betonquest.compatibility.redischat;
 
 import dev.unnm3d.redischat.api.RedisChatAPI;
+import net.kyori.adventure.text.Component;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
@@ -53,6 +54,11 @@ public class RedisChatInterceptor implements Interceptor {
     @Override
     public void sendMessage(final BaseComponent... message) {
         player.spigot().sendMessage(message);
+    }
+
+    @Override
+    public void sendMessage(final Component message) {
+        player.sendMessage(message);
     }
 
     @Override