+++ src/main/java/org/betonquest/betonquest/notify/ChatNotifyIO.java
@@ -1,5 +1,6 @@
 package org.betonquest.betonquest.notify;
 
+import net.kyori.adventure.text.Component;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.QuestException;
@@ -16,6 +17,16 @@ public class ChatNotifyIO extends NotifyIO {
 
     @Override
     protected void notifyPlayer(final String message, final OnlineProfile onlineProfile) {
+        final Conversation conversation = Conversation.getConversation(onlineProfile);
+        if (conversation == null || conversation.getInterceptor() == null) {
+            onlineProfile.getPlayer().sendMessage(message);
+        } else {
+            conversation.getInterceptor().sendMessage(message);
+        }
+    }
+
+    @Override
+    protected void notifyPlayer(final Component message, final OnlineProfile onlineProfile) {
         final Conversation conversation = Conversation.getConversation(onlineProfile);
         if (conversation == null || conversation.getInterceptor() == null) {
             onlineProfile.getPlayer().sendMessage(message);